web-interface/requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycups==2.0.4
requests==2.32.3
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
